update to mstest 3

+++ src/Test/Desktop/OpenRiaServices.Common.Test/ExceptionHelper.cs
@@ -255,7 +255,7 @@ namespace OpenRiaServices.Client.Test
             Assert.AreEqual(message, e.Message);
 
             HttpWebResponse response = e.Response as HttpWebResponse;
-            Assert.AreEqual(errorCode, response.StatusCode);
+            Assert.AreEqual(errorCode, (int)response.StatusCode);
             return e;
         }
